Build of control sheet
Haven't created every combination of column names yet, or made it so that variable 1 is always filled in (e.g. this is still an issue: var_1 = NA, var_2 = "Age").

Filtered available columns so that only the core disaggregations (found in multiple indicators) are kept
Have built the control sheet so that var_1 corresponds to level_1 and has an associated var_group_1.


Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>