control sheet including levels
worked out how to get all possible variable combination orders,
but have come to the conclusion that including levels in the
initial user selections is probably not necessary, as we will
usually care about all levels when looking at interactions.

By simplifying the selection process, we can massively cut down
the size of the control sheet. I am not sure that a control
sheet with levels included will work as it would have to be so large
to include all possible permutations, and even if it was possible I
suspect it would be quite slow.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>