sorting difficult filtering issues
initially thought I would use identify_All_vs_NA.R, which was originally
control_sheet_including_all_levels.R, but this wasn't useful.

have started to sort out filtering when one of the selected columns is
entriely nested in an unselected column.

It is possible some of the issues I am facing are because I have made impossible selections in the config

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>