use shinywidget to solve reactivity issues
We were struggling to get an output unless the user used all possible selection.
This function does all the hard stuff for us but seems too easy -
can we build the next steps on top? i.e. can we then add a select indicator dropdown
and save the option chosen?

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>