Filtering script completed?
Have made it so that if a multiple user-selected variables are nested in
other unselected variables, the filtering works correctly. E.g if the
user selected region and ethnicity, but these were only available for
'England', and 'England & Wales' respectively, and Country was not
selected for display by the user, the filtering would still work.

Needs more testing. May break when it is a straightgforward indicator
with no nesting (might need another if statement).

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>