filtering completed
Added in a new output so that information is available to create dropdowns to appear
for unselected disaggregations that don't have a headline value.

Currently filtering is over-engineered and I think it does more than it needs to.
I forgot that selected disaggs will always be layered (interacting) and
made it so filtering worked with non-interacting disaggs as well.

Still needs a bit of cleaning but has passed tests.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>