filtering and plotting completed
tidied up the folders so that only relevant scripts are in the main folder. The rest were moved to Archive.
They can probably be deleted.
Changed the name of some of the files so that they make more sense.
Changed config to user_selections as this is a better description.
User_selections will not be needed in the final script.

This should now be ready to integrate with the rest of the app.

Some testing done, but not extensive.

Signed-off-by: EmmaWoodONS <Emma.Wood@ons.gov.uk>